+++ composer.json
@@ -74,6 +74,7 @@
     "require-dev": {
         "ext-simplexml": "*",
         "brianium/paratest": "^6.3",
+        "composer/semver": "^3.3",
         "fidry/makefile": "^0.2.0",
         "helmich/phpunit-json-assert": "^3.0",
         "phpspec/prophecy-phpunit": "^2.0",
